move the CSCCondDB plugins in the plugin directory
